fonts: Libre Franklin, in four weights, for the four the game asks for
ergoec, arialnb, anewhope and ocr_a now resolve to weights of Libre Franklin
(OFL 1.1, and its Cyrillic is why a TrueType face is involved at all - the
shipped fonts have 256 glyphs and not one of them is Cyrillic). The faces are
in assets/fonts/src; the atlases beside them are generated by
tools/fontgen/build_fonts.py and checked in.

Checked in rather than built, because building them would put freetype, pillow
and numpy between someone and a working game for the sake of four files that
change when a typeface changes, which is roughly never. That trade only works
if they cannot go stale, so build_fonts.py --check rebuilds them into a
temporary directory and compares, and that runs in the tests stage and in CI.

The size the atlas is generated at and the size the game lays out with are now
two separate numbers, which is the substantive change here. They used to be one:
--size was texels per em AND the reported point size, so asking for a better
atlas doubled the height of every menu. A 48 texel em reported at 16 point is a
sharp 16 pixel font. That also means glyph placement stops being integral - a
glyph 41 texels wide in a 48 texel em is 13.67 points, and rounding that to 14
is a two per cent error in the size of every letter - so the four placement
fields in .jkxfont are floats.

The point sizes in the manifest are the shipped ones, but they decide the whole
vertical layout of the interface, so build_fonts.py --assets <base> will read
them out of the retail .fontdat files rather than trust the manifest. Advances
are deliberately not matched: the text is set in Libre Franklin's own metrics,
which was the decision, and it reflows.

aurabesh is not in the manifest and is not replaced. It is an invented alphabet
rather than a typeface, and a Latin face standing in for it turns alien signage
into readable English. The generator's bitmap path exists for it; it needs the
game installed, and the command is in build_fonts.py's docstring.

+++ code/rd-vulkan/tr_font_sdf.cpp
@@ -63,11 +63,16 @@ typedef struct {
 	unsigned int	glyphCount;
 } jkxFontHeader_t;
 
+// The atlas rectangle is in texels and integral; the placement is in the units
+// the game lays out with and is not. The field can be generated at 48 texels
+// per em and reported at 16 point - that is what keeps atlas resolution a free
+// choice - and at that ratio a glyph 41 texels wide is 13.67 points, which does
+// not survive being an integer.
 typedef struct {
 	unsigned int	codePoint;
 	unsigned short	x, y, w, h;			// where in the atlas
-	short			xoff, baseline;		// where against the pen, in game units
-	short			gw, gh;				// and how big it is drawn, in game units
+	float			xoff, baseline;		// where against the pen
+	float			gw, gh;				// and how big it is drawn
 	float			advance;
 } jkxFontGlyph_t;
